rename keys file
