+++ src/checkbox.tsx
@@ -1,6 +1,6 @@
 'use client';
 import * as React from 'react';
-import {SPACE} from './utils/key-codes';
+import {SPACE} from './utils/keys';
 import {useControlProps} from './form-context';
 import Inline from './inline';
 import {Text3} from './text';
