+++ src/radio-button.tsx
@@ -1,6 +1,6 @@
 'use client';
 import * as React from 'react';
-import {SPACE, LEFT, UP, DOWN, RIGHT} from './utils/key-codes';
+import {SPACE, LEFT, UP, DOWN, RIGHT} from './utils/keys';
 import {useControlProps} from './form-context';
 import {combineRefs} from './utils/common';
 import {Text3} from './text';
