+++ src/touchable.tsx
@@ -5,7 +5,7 @@ import ScreenReaderOnly from './screen-reader-only';
 import * as classes from './touchable.css';
 import {useTheme} from './hooks';
 import {isInsideNovumNativeApp} from './utils/platform';
-import {ENTER, SPACE} from './utils/key-codes';
+import {ENTER, SPACE} from './utils/keys';
 import {getPrefixedDataAttributes} from './utils/dom';
 
 import type {DataAttributes, TrackingEvent} from './utils/types';
